+++ CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added POINTING WEAPON (by officer) table type
 - Added data loader to combine multiple files that span a single year into a single dataset
-- Support for more text date column formats in Arcgis loader.
+- Added support for more text date column formats in Arcgis loader.
+- Added url_contains input to get_count, load_iter, load, and load_from_csv of Source class to distinguish between multiple datasets matching a data request
+- Added datasets input to get_years to allow getting the years in specific datasets.
 ### Changed
 ### Deprecated
 ### Removed